refactor(event-stream): extract control-frame state machine to control.rs
Pure code motion for #6235. Move the daemon->helper control-frame decode
and the drain/resync state machine out of mod.rs into a dedicated
control.rs submodule: process_control_frames (payload-length ceiling
#2879, ACK-window validation #2959, Pause/Resume, DrainRequest dispatch)
and handle_drain_request (the DORMANT seq-fenced drain #2876/#2882 with
the #2875 pause-poison FullResync path).

Moved verbatim (no logic change). Both functions are widened from private
to pub(super) so the connected loop and the control_frames / drain test
suites resolve them via the re-import. The codec MSG_* / FRAME_HEADER_SIZE
imports are annotated #[allow(unused_imports)] since, after this move,
they resolve for the control (and soon connection) submodules via
`use super::*` and for the test tree, not for mod.rs directly. cargo
build green.
